+++ R/R/roll.R
@@ -2,7 +2,7 @@
 #'
 #' @param x \code{xts} object with a column named \code{Close}, representing closing prices.
 #' @param width integer width of the rolling window to use, or vector of endpoints defining the intervals to use.
-#' @param signed a \code{logical} value indicating whether non-positive estimates should be preceded by the negative sign instead of being imputed. Default \code{FALSE}.
+#' @param signed a \code{logical} value indicating whether signed estimates should be returned.
 #' @param na.rm a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds. Default \code{FALSE}.
 #'
 #' @return Time series of spread estimates.
@@ -20,16 +20,19 @@ ROLL <- function(x, width = nrow(x), signed = FALSE, na.rm = FALSE){
   R2 <- lag(R1, 1)
 
   # drop leading NA
-  R1 <- R1[-1]
+  R1 <- R1[-c(1:2)]
   R2 <- R2[-c(1:2)]
 
   # expectations
-  E1 <- rmean(R1, width = width-1, na.rm = na.rm)
+  E1 <- rmean(R1, width = width-2, na.rm = na.rm)
   E2 <- rmean(R2, width = width-2, na.rm = na.rm)
   E12 <- rmean(R1*R2, width = width-2, na.rm = na.rm)
 
+  # Number of observations
+  N <- rsum(!is.na(R2), width = width-2)
+  
   # squared spread
-  S2 <- 4 * (E1*E2-E12)
+  S2 <- N / (N-1) * 4 * (E1*E2-E12)
   
   # square root
   S <- sign(S2) * sqrt(abs(S2))